Parser can now be stopped before a certain tag.
Added methods to the dcmdata parser that allow the read process
to be stopped before a certain tag is fully processed. This also
works if the given tag (which must be on the main dataset level)
is not present in file, i.e. no higher tag is parsed/processed.
Added --stop-before-elem command line option to dcmdump to
enable use of these new methods.

This closes DCMTK patch #347. Thanks to Jeroen Eggermont for
the suggestion and the (partial) patch.