Fixed installing manpages with CMake.
CMake now installs manpages from the source code directory unless a newer
version of a manpage exists in the build directory. This fixes installing
manpages in out-of-source builds in case they are not explicitly generated
via 'make man'.
Inside source builds are not affected by this modification, since the manpage
in the build and source directory are always identical in that case.
Futhermore, CMake now omits unwanted directory index manpages generated by
doxygen.