Fixed linking some dcmsr unit tests with MSVC.
Fixed linking dcmsr_addItems and dcmsr_getItems when building shared libraries
(DLLs) with Visual Studio. The tests had previously been disabled when
building shared libraries since Visual Studio is unable to handle some of
the symbols that are part of the DSRListOfItems template.
This workaround employs a dummy class to create specific symbols for the tests
only.

Closes DCMTK bug #694.