Fixed CTest failing tests that did not compile.
The previously introduced script based approach to run the unit tests for
allowing additional parameters (e.g. -x) caused unit tests to fail if the
respective test binary was unavailable instead of the correct and previous
behavior to skip them.
The respective binaries are now correctly marked as unit test dependency to
restore the previous behavior.