Fixed a mistake in the code generating arith.h.
The configuration test for inspecting the fundamental arithmetic types was
written under the wrong assumption that std::isnormal tests for non-subnormal
values. Instead isnormal returns true if the value is neither zero, subnormal
NaN or infinity.
The test now compares the given value with FLT_MIN resp. DBL_MIN instead,
which should be more appropriate. Luckily, the results seem to be the same
on the platforms we have tested so far.