Check override keys in img2dcm library code.
img2dcm silently ignored errors if override keys defined with the --key
option (or by the library user) can neither be parsed to a valid tag nor
tag dictionary name. Now the error is handled and returned to the user.

Additionally, the related method now uses a unique pointer to ensure that
all code paths are cleaning up memory in case of error.

This closes DCMTK bug #589.