Refactored cross compiling support with MinGW/Wine.
When cross compiling for Windows using MinGW/Wine, the configuration tests
will now run in a specifically prepared wineprefix that is configured for
linking the detected/configured MinGW runtime DLLs.
The Wine instances will also be spawned as detached processes, with their
IO streams redirected to temporary files. This works around a problem where
CMake was waiting for the whole 'wineserver' process to shutdown instead of
only the spawned application, before continuing the execution. The temporary
files will be read and displayed etc. such that the user experience does not
change significantly compared to before.
This commit will break running the unit tests via Wine, which will be
addressed by a follow-up refactoring of the unit test related CMake
configuration.