Warn if patient name may lead to broken dir name.
The option --sort-on-patientname uses the patient name from received files
in order to create a storage directory. The dir name is always in ASCII.
storescp already defines a mapping from Latin-1 to ASCII. However, if
a patient name uses a different character set, the latin-1 mapping is
silently performed anyway.

The new code does print a warning for cases where the patient name is not
in ASCII or latin-1 denoting that the resulting directory name may be broken.
Also, the current behaviour is documented.

This fixes DCMTK Bug #557.