Fixed and refactored running the unit tests via CTest.
CTest will now always set the environment variable DCMDICTPATH appropriately
such that the DICOM dictionaries from the source code folder will be used
for running the tests. This prevents the test from failing caused by
accidental use of a wrong dictionary or failing to load the dictionary.
The test can now be executed via Wine and utilizing the Android emulator
again, in both cases now also supporting to run the test in exhaustive mode
(via make test-exhaustive etc.) and also correctly setting DCMDICTPATH.
For both Android and Wine a CTestCustom.cmake file will be created to ensure:
The Android emulator will be started, prepared and stopped as necessary, that
is, the unit tests, required shared objects and now also the DICOM
dictionaries will be copied to the device before running the tests.
DCMDICTPATH and LD_LIBRARY_PATH will be set accordingly on the device during
test execution.
The Wine variant of the CTestCustom.cmake will ensure that the 'wineserver'
process is already running during test execution, preventing CTest from
attaching itself to that process leading to unnecessary and annoying wait
times before executing the next test.