Ensure SPS/RP Description/Code is never empty.
Per DICOM standard, either SPS Description or Scheduled Protocol Code, as well
as either Requested Procedure Description or Requested Protocol Code must be
filled in (type 1C, with the condition that either one of both attributes in
each of both cases must be supported). The old code permitted that if the
respective sequence is filled in, the description may have an empty value.

Now an empty value is forbidden for the descriptions and for the sequences
in any case. One of both (for each case)  has to exist with a (non-empty)
value.

Thanks to Anne-Gaelle Berge <anne-gaelle.berge@kereval.com> for reporting
a related issue.

This closes DCMTK Bug #472.