Enhanced XSD definition of "UniqueIdentifier".
Enhanced XML Schema (XSD) definition of the DICOM value representation
"Unique Identifier" (UI), which was not as strict as possible. Actually,
the first component of an OSI Object Identification (numeric form) always
starts with "0", "1" or "2" but this is currently not checked in order to
accept values that are sometimes used for testing purposes (e.g. "99.xxx").

Thanks to Mathieu Malaterre <mathieu.malaterre@gmail.com> for the report
and suggested fix.