Updated COPYRIGHT file.
The COPYRIGHT file now also includes the license notices of the
external libraries that DCMTK can optionally be configured to
use, such as zlib, libxml2 and openssl. This is to make sure
that the copyright notices remain available to users that only
download and install the executable binary package, which does
include the COPYRIGHT file.

This closes DCMTK bug #384.