Now setting LossyImageCompression when decompressing.
All DCMTK-based image decompression codecs now insert the attribute
LossyImageCompression with a value of 01 when decompressing an image
encoded in a lossy transfer syntax. This makes sure that a warning
about the reduced image quality remains in the dataset even if the
encoder did not appropriately create it.

This closes DCMTK feature #307.