Partially refactored 'errno' usage within dcmnet.
Added specialized error code handling for WinSock by using
OFStandard::getLastNetworkErrorCode() and the respective WinSock error
constants (WSAEINTR) as appropriate within dulfsm.cc.
Most of the other files in dcmnet and the rest of DCMTK already have a
separate Windows implementation in this regard, so this fixes most of the
remaining problems.