Converting large attributes to UN in explicit VR.
When writing in explicit VR, some VRs are encoded with a 2-byte length field,
whereas other VRs are encoded with a 4-byte length field. When writing
implicit VR, all length fields are always encoded with 4 bytes. In some
cases attribute values of the affected VRs may be larger than 64 kBytes,
in which case the dataset can be represented in implicit VR, but not in
explicit VR. Instead of refusing to write such datasets in explicit VR,
DCMTK now automatically converts the affected attributes to UN
(or, when generation of post-1993 VRs is disabled, to OB).

This implements DICOM CP 1066 and closes DCMTK issue #331.