First check number of components before value.
compare() was changed to first check the number of values of an element
instead of first comparing the values. The reason is that it takes
significantly more effort to compare the values which is often not
necessary if the number of components is different anyway. This is
especially useful for large binary values (like pixel data). The
following commit will introduced specific comparsion operators for such
elements.

This might result in calls to compare() that now return -1 that earlier
returned 1, and vice versa. This is not considered harmful since the
functionality is quite new in DCMTK, and the majority of use cases would
only check whether two elements are equal (0 returned) or not. The new
implementation will never return equality (0) where it earlier returned
-1 or 1.

Acknowledgement: This work has been supported in part by the "QIICR" project.