Fixed possible role selection config. problem.
Fixed possible role selection problem for association acceptors. If the
first presentation context added to the association configuration used a
default role, any other presentation context with role selection
configured would not configure correctly. The reason is that when
creating the underlying association profile in the first step, one
already has to know the role selection profile, which does not exist if
not instantly added together with the association configuration profile
(together with the first presentation context). This commit ensures that
an empty role selection profile can and is created in any case and all
presentation contexts with role selection being added refer to that profile.