Opt. accept Default role if SCP role expected.
Added option to dcmnet assocation acceptor code that allows to accept
a Default role (no role selection in request) even if the acceptor is
configured to only accept SCP role. This is helpful for faulty clients
in services like Storage Commitment or Retrieve (via C-GET) very
association requestors have to use role selection if they want to be
correct with respect to the standard. Sine some are not, this option
makes sure that such requests can be accepted (with return Default role,
too).

Added option to association acceptance code in dcmnet in order to allow
for accepting