Added a partial workaround for DCMTK bug #758.
Added a check to the configuration and unit tests of std::vector/OFVector to
ensure that the construction of recursive data structures based on OFVector
does not cause a stack overflow.
Modified the default constructor of DCMTK's OFVector implementation to no
longer predictively construct some elements, such that it passes the new test.