Fixed using wrong build type for configuration test.
CMake will now use the same build configuration for running the unit tests
(e.g. 'Release' or 'Debug') as the one that was chosen for building DCMTK.
This fixes different configuration results incompatible with the later build,
e.g. not detecting the problems with Visual Studio 2005's STL vector that
led to failing the OFVector unit test.