Many workarounds for SunPro.
Many workarounds and some actually meaningful fixes for the SunPro compiler.
In general the compiler at best seems to have a basic implementation of
some parts of the C++ type system but does not actually implement it, e.g.
fails to understand more complex types and perform any meaningful operations
on them.
The (perhaps) meaningful fixes are:
1. The configuration test for detecting whether the compiler is able to
identify default constructable classes via SFINAE now also tests whether the
result is the expected one and not only whether the code compiles.
2. The configuration tests for std::list and std::vector now also tests
whether a const_iterator can be compared to a mutable iterator of the
respective template, e.g. whether the compiler understands that a
const_iterator can be constructed out of an iterator for the comparison on the
fly.
3. The configuration tests for 'getgrnam_r' and 'getpwnam_r' now also ensure
that both functions take the expected number (and types) of arguments.