Fixed wrong usage of alignas.
Fixed using alignas within typedefs (which was first introduced to realize
compatibility with the implementation for Visual Studio). The new
implementations of OFalign and OFalign_typename for alignas only define the
given type without the alignment (for handling arrays) and add the alignment
modifier on preprocessor level.
Adjusted the documentation of OFalign accordingly and removed the now
unnecessary configuration test whether type definitions using alignas are
possible (since even if the test succeeded the compilers most likely didn't
actually align the resulting type as one would expect).

Thanks to Stefan Hengelein and Peter Klotz for reporting the issue.