Added a workaround for C++11 compiling errors.
Added a workaround for errors when enabling C++11 (DCMTK_ENABLE_CXX11) without
enabling std::tuple. It is of course debatable whether this scenario makes any
sense at all, but since it is currently the default behavior this workaround
should increase usability.
We should possible override the user choice whether or not to enable
std::tuple in case C++11 was enabled in the future or provide a full C++11
implementation of OFtuple (is there any C++11 compiler that does not have
std::tuple?).