Added support for multiple finding sites (DCMSR).
Added support for multiple finding sites in TID 1411 (Row 2) as introduced
with CP-1591 (Allow multiple finding sites for single regions of interest
in measurement templates and segmentations).

Please note that this commit also changes the name of the corresponding
method in class TID1411_VolumetricROIMeasurements from setFindingSite() to
addFindingSite(). This name change makes sure that the user of this class
is warned (if he/she called the "old" method in his source code).

Acknowledgement: This work has been supported in part by the "QIICR" project.