Fixed bug when importing Series and Frame of Ref.
Fixed bug when importing Series and Frame of Reference, caused by a
wrong "forwarding" import() call that used the parameters in the wrong
order. This behaviour was hidden by another issue where the same
parameters have been switched around (compared to the header) in the
calling method's parameter list. When the latter issue was fixed
earlier (see commit a9b866) the former issue got revealed, leading to
problems when both parameters (readSeries and readFoR) were called with
different values.

This commit fixes the order in the import() call, and makes the
parameter order in all import() methods consistent.

CAUTION: In order to make the user aware of the critical parameter order
change, the name of the import() method is now importHierarchy() and a
prominent note was added that the orders of parmeters is incompatible
to the old one.