Add extra content items to extensible templates.
Added new method that allows for adding extra content items to extensible
SR templates. Also added new test case and enhanced existing ones to verify
that it works as expected.

Renamed existing method insertTemplate() to insertExtraTemplate() in order
to be consistent with the new method addExtraContentItem(). This method was
only used in two test cases so far.

Acknowledgement: This work has been supported in part by the "QIICR" project.