Added comparison operators to various SR classes.
Added "equal" and "not equal" comparison operators to all document tree node
and underlying value classes as well as to the wrapper class for "SR content
items". These operators will be needed for an upcoming enhancement of the
TID1600_ImageLibrary class, which is part of the "content mapping resource"
sub-module "dcmsr/cmr".

Acknowledgement: This work has been supported in part by the "QIICR" project.