Consistently use Command Priority "MEDIUM".
Consistently use the default value "MEDIUM" (0x00) for Priority (0000,0700)
in all DIMSE request messages, i.e. for C-STORE, C-FIND, C-GET and C-MOVE.
Before, some network tools and also the DcmSCU class used "LOW" (0x02) for
unknown reasons.

The Priority value is still "hard coded" (i.e. not configurable) since we
are not aware of any DICOM implementation that would make a difference on
the value of (0000,0700).