Use getnameinfo() for reverse DNS lookups.
Added a new function OFStandard::getHostnameByAddress() that performs
a reverse DNS lookup of a hostname based on an IP address. This function
uses the protocol independent getnameinfo(3) function if available instead
of gethostbyname() or gethostbyname_r().
Removed function OFStandard::getHostByAddr(), which due to its interface
cannot be ported to use getnameinfo(3) and which is not used anymore
in the toolkit.

This closes DCMTK bug #715.