Make sure the calling host is being checked.
DcmSCP has a virtual method in order to check whether the calling host
should be accepted or not, so that derived classes could implement their
own acceptance policy for hosts. However, it turned out that this method
has not been called at all so far. Now the host is checked when
connecting.

Thanks to Domen Soklic <domen.soklic@cosylab.com> for the report.