Refactored/Fixed (Ident.) Pixel Value Transf. FG.
Refactored and fixed the (Identity) Pixel Value Transformation Functional
Group. The Identity version of the Functional Group sets fixed values for
Rescale Slope/Intercept/Type, and thus is a specialization of the regular
Pixel Value Transformation Functional Group. In order to use the existing
Identity version for both type of FGs, the code for the Identity version
has been generalized.

The original trigger for the fix was that it was undecidable for the FG
factory code to decide by the FG sequence tag which of both FGs to
instantiate (reflected by a faulty doubled if condition). Also, the file
has been renamed, so that name the reflects the more general Pixel Value
Transformation FG.

Fixed misleading code indentation reported by gcc.

Thanks to Martin Strunz for the report.