Updated "dcmrt" classes based on DICOM 2017e.
Updated automatically generated IOD and sequence C++ classes for the various
RT objects based on the current edition of the DICOM standard (2017e).

Also fixed three attribute definitions that are used in the RT Beams Module
and that were marked as "retired". This was apparently true some years ago
but not anymore according to the current edition of the DICOM standard.