Removed double definition of retired attributes.
Removed double definition of retired DICOM attributes, i.e. those that can be
extracted automatically from part 6 and 7 of the DICOM standard are no longer
added manually at the end of the dictionary file.

This also fixes the status of Destination AE (2100,0140), which had been
retired in the past but this retirement has been withdrawn (apparently).