Fixed possibly wrong calculation of value count.
The method getVM() cannot be used for VR classes such as FL, FD and UL to
determine the number of stored values because derived classes (i.e. OF, OD
and OL) always return 1 (since this is how the Value Multiplicity is defined
for these Value Representations). Also enhanced the comment explaining this.