Fixed issue with uninitialized memory buffer.
Fixed an issue with an uninitialized memory buffer when processing
monochrome DICOM images with invalid data element values in the Image
Pixel Module. For example, when the value of BitsAllocated was too high,
parts of the memory buffer that stores the intermediate representation
of the pixel data did not get initialized (even though it should have
been "zeroed").

Thanks to GwanYeong Kim <gy741.kim@gmail.com> for the original report.