Enhanced the way the unit tests are executed (CMake).
The unit tests are now added using add_test(NAME ...) instead of the old
add_test() command. This enables using generator expressions to dynamically
resolve the location of the test executables.
Also ensured that the test-exhaustive targets pass $<CONFIG>, which is
required for setups that use this mechanism (e.g. Visual Studio).

Closes DCMTK bug #783.