Added --extract-xml-single option to findscu.
Added --extract-xml-single option to findscu and enhanced the underlying C++
class accordingly. This new option allows for extracting all C-FIND-Response
datasets of an association to a single XML file. If character set conversion
is available, all datasets are converted to UTF-8 (in order to avoid issues
when different character sets are used). See documentation for details.