Use default timeout for first PDU to be read.
In DUL_RequestAssociation(), i.e. when requesting an association, use the
default timeout, the one that is passed to ASC_initializeNetwork(), for the
first PDU to be read (if no other timeout value is defined). That means, the
behavior of this function is now consistent with DUL_ReceiveAssociationRQ().
For the various command line tools, this timeout value is usually specified
by the --acse-timeout option. So, it now also works with echoscu, findscu,
storescu, dcmsend, etc.

Also added output to TRACE logger when setting the request/receive timeout.