Fix cmake error on Linux with DCMTK_ENABLE_CXX11.
Configuring on Linux with DCMTK_ENABLE_CXX11 enabled generates an error
in the configuration step:

    CMake Error at CMake/GenerateDCMTKConfigure.cmake:1252 (MESSAGE):
      ...
      /path/to/dcmtk/config/tests/../math.cc:88:12: error:
      '::isnan' has not been declared

           return ::isnan(f);

One way around this is to ensure that the preferred std::isnan()
function is used instead. This commit fixes the following error in the check
for that function by specifying the double overload:

    FAILED: CMakeFiles/cmTC_62c88.dir/CheckSymbolExists.cxx.o
    /path/to/dcmtk-build/CMakeTmp/CheckSymbolExists.cxx:29:11: error:
statement cannot resolve address of overloaded function
     std::isnan;return 0;

Thanks to Max Smolens (github user msmolens) for the report and patch.