Added import() method for FoR and FoR checks.
Added import() method to import Patient, Study and FoR information from
a given source image but, unlike the related DcmIODCommon::import()
methods, do not require knowledge whether Frame of Reference should be
imported or not.

Also added FoR-related checks when segmentation objects are written,
i.e: 1) Check whether FG is set on segmentation if Derivation Image FG is
missing and 2) Check whether Pixel Meausures FG, Plane Position (Patient)
FG as well as Plane Orientation (Patient) FG are present if FoR is set.