Fixed two bugs in DNS lookup code.
Fixed a memory leak in OFStandard::getAddressByHostname() caused by a
missing call to freeaddrinfo(). Added counter to limit the number of
retries in this function and in OFStandard::getHostnameByAddress()
in case the network subsystem returns EAI_AGAIN.

Thanks to Peter Klotz <Peter.Klotz@ith-icoserve.com> for the bug report
and patch.