Added new method getNumberOfValues().
Added method getNumberOfValues() to all VR and other classes that are
derived from DcmObject (such as DcmItem). This new method always returns
the number of values currently stored in the corresponding object/element,
in contrast to getVM(), which sometimes returns the constant 1, or card(),
which is not available for all classes.

This closes DCMTK Feature #808.