Fix posible crash if network data is malformed.
If an unknown's item length field was larger than the available amount
of data, parseDummy() would detect this error and report it, but the
caller would ignore this and continue anyway.

The following OFStandard::safeSubtract() would not catch the error with
carefully chosen values: parseDummy() adds 4 to the two-byte length
field that is read from the package. This allows to construct a value
that does not fit into unsigned short. Thus, the pointer buf gets
incremented by 65538 while userLength would only get the value three
subtracted.

This commit fixes both of the above, but either one would be enough to
fix the issue.

Thanks to Uli Schlachter <psychon@znc.in> for the report and patch.