Fix underflow in networking code.
The amount of data in a PDV is the PDV's length field minus two, since
there are two bytes of other data. However, the code did not check that
the calculation pdvLength-2 did not underflow. This results in passing
incorrect huge buffer sizes to other functions later on which thus read
beyond the end of the buffer.

Thanks to Uli Schlachter<psychon@znc.inc> for the report and patch.