Fix possible bypass of length check.
For huge elements, it is possible that (int)elem->getLength() overflows.
For example, an element with length 2164260863 is interpreted as
-2130706433. This allows to wrongly pass the length check.

Fix this by moving the check to the other size and casting maxlen
instead. This is safe since the maxlen value is an integer constant in
all callers.

Thanks to Uli Schlachter<psychon@znc.in> for the report and patch.