Use POLLOUT if connect() returns EINPROGRESS.
Since lately DCMTK networking uses poll() if available to handle network
connections. When starting a connection() using connect, it may return
EINPROGRESS indicating to try again later using poll() to write data to
the socket. Writing though requires the POLLOUT flag in poll(), not
POLLIN as it has been used so far.

Thanks to github user sercxjo for the report and patch.