Added a missing test for universal matching.
Added a missing test for universal matching before performing combined
matching. The missing test was so far not a bug since the only existing
combination is for combined date time matching, in which case the only
possible universal match is an empty date or time.
This was already accounted for in the actual matching implementation, so this
commit is only for preventing bad surprises in the future where other types of
matching combinations might be defined.