Introduced upper size limit for A-ASSOCIATE PDUs.
Introduced upper size limit for incoming A-ASSOCIATE-RQ and -AC PDUs
to avoid denial of service through malformed PDUs causing the network
layer to allocate up to 4 GByte of memory. The default limit is 1 MByte.
This limit can be modified or disabled through the variable
dcmAssociatePDUSizeLimit.

Thanks to Klaus Eisentraut <klaus.eisentraut@siemens-healthineers.com>
for the bug report.

This closes bug #825.