Fixed issue with incorrect rendering of overlays.
Fixed issue with incorrect rendering of overlay planes embedded in the pixel
data when performing rotation and scaling on the same DicomImage instance,
e.g. when calling dcm2pnm with options "+Sxf 0.5 +Rl" or "+Sxf 1.5 +Rtd".

Also enhanced output to trace logger in order to discover the reason for this
issue (and possibly similar issues in the future).

This closes DCMTK Bug #828.