Class OFReadWriteLock now uses SRW locks on Windows.
The Win32 implementation of class OFReadWriteLock class now uses
Slim Reader/Writer (SRW) Locks, which are available since Windows Vista,
instead of the older implementation based on a Mutex, a Semaphore and
a counter, because SRW locks are much faster. Unfortunately SRW locks
require different functions for unlocking read and write locks, which
causes an API change in class OFReadWriteLock and in class
GlobalDcmDataDictionary (which internally uses a read/write lock to
protect access to the DICOM dictionary.)

Thanks to Christian Fremgen <Christian.Fremgen@ith-icoserve.com>
for the patch that has inspired this implementation.

This closes DCMTK Patch #810.