Fixed incompatibility between FindIconv versions.
In version 3.11 CMake introduced its own FindIconv.cmake, which sets
ICONV_FOUND instead of LIBICONV_FOUND, as it is done by the DCMTK version
FindICONV.cmake. The other variable names also omitted the "LIB" part.
Furthermore, the CMake version handles detection of extern versions of
LibIconv and the built-in version in the C standard library differently. This
should now also be adjusted.
Additionally the CMake command "find_package" threats found modules as
case-sensitive, so the DCMTK version had to be renamed to match the new
built-in version.

This closes DCMTK Bug #837.