Fixed bug on responseCounter of findSCU.
The responseCounter of findSCU did not increment properly and therefore the
file rsp001.dcm was overwritten every time. By passing the counter as a
reference, the files are now named correctly an not overwritten. The findUser
function was also adapted in dcmqrtis and dimse and properly documented.

This fixes bug 809.