Avoid type mismatch warnings (sprintf and sscanf).
Avoid type mismatch warnings on some platforms when using sprintf() and
sscanf() with integer variables of a specific size (number of bits).
This is done by using the format macro constants defined in "inttypes.h".
For reasons of supporting older compilers than C99 and systems that do not
provide this header file (with the macros defined), the old code is still
retained.

This closes DCMTK Bug #764.