Implemented DICOM CP 1653 in the JPEG codec.
Implemented DICOM CP 1653, which forbids the use of the YBR_FULL
photometric interpretation in lossy JPEG transfer syntaxes, which was
the default in DCMTK. Only YBR_FULL_422 (with 4:2:2 subsampling) is
permitted now. Changed the default behaviour of the JPEG encoder
accordingly and renamed dcmcjpeg's --sample-422 command line option to
--nonstd-444. Furthermore, subsampling is now suppressed in all lossless
JPEG encoding processes.

This closes DCMTK issues #716, #717 and #811.