Fixed wrong documentation of calcElementLength().
The documentation of DcmObject::calcElementLength() stated that the function
never returned an undefined length. This is no longer true since a previous
commit introduced overflow handling that maps an overflow during length
calculation to the maximum possible value (clamping), which, coincidentally,
equals the constant defined as DCM_UndefinedLength.
This commit adjusts the documentation of calcElementLength() in DcmObject and
derived classes accordingly.