Updated DIMSE compatibility flag.
Updated DIMSE compatibility flag to reflect recent changes in the networking
code. Windows now uses socklen_t as it originally should, due to the
respective configuration test having being fixed recently. This normally
should not affect the behavior of the network code (as socklen_t should
typically be defined to the type that was used before anyway), but better safe
than sorry.