Fixed incorrect assumption in OVector's unit test.
Since both begin() and end() return the same iterator for an empty vector,
it is not obvious whether such an iterator shall refer to begin() or to end()
after the first insert operation.
OFVector's unit test so far assumed it would refer to begin(), but, this is
not always the case, leading to the test failing on Visual Studio 2017 when
compiled in Debug mode.

This closes DCMTK issue #859.

Thanks to Bengt Gustafsson <Bengt.Gustafsson@contextvision.se> for the report
and detailed analysis of the issue.