Create list of elliptic curves for TLS dynamically.
The list of elliptic curves to be negotiated as part of a TLS
connections was so far hard-coded in DCMTK, based on the elliptic curves
with at least 256 bits that are supported in a default OpenSSL 1.0.2
build (the oldest supported version). However, some distributions (e.g.
RHEL) ship with modified OpenSSL libraries that further restrict the
list of supported curves, causing an error in the dcmtls module. We now
test at runtime which elliptic curves are supported and only negotiate,
thus avoiding this problem.

Thanks to Peter Klotz <Peter.Klotz@ith-icoserve.com> for the bug report.

This closes DCMTK issue #863.