Fixed memory leak in multithreaded apps on Win32.
Fixed a memory leak occuring in multi-threaded applications on Windows
when logging output was performed in multiple threads and the user did
not manually call dcmtk::log4cplus::threadCleanup() when terminating a
thread in order to clean up the thread-local storage allocated by oflog.
Instead of thread local storage, oflog now uses fiber local storage and
registers a callback function that automatically cleans up the memory
when a thread ends, in a similar manner to the Posix pthreads code,
which never required the used to manually call a cleanup function. The
previous behaviour can be reactivated by defining the macro
DCMTK_LOG4CPLUS_AVOID_WIN32_FLS.

Thanks to Chinna Durai <chinnadurai410@gmail.com> for the bug report
and sample code to demonstrate the problem.