Disable TLS 1.3 for the historic security profiles.
When compiling with OpenSSL 1.1.1 or newer, support for the TLS 1.3 protocol
is available. TLS 1.3 uses a completely different set of ciphersuites and
a different set of APIs in OpenSSL. When one of the historic TLS security
profiles (3DES, AES, and NULL) is selected, we set the maximum permitted TLS
protocol version to 1.2, because otherwise the TLS 1.3 support would lead
to unexpected behaviour (e.g. encrypted transmission despite the NULL profile
being active).

This closes DCMTK issue #865.