Added initial support for "SCT" code definitions.
With CP-1850 the use of the SRT coding scheme has been deprecated. As a first
step towards the migration of "SRT" to "SCT" (SNOMED CT) codes, the subset of
codes that are defined for the SRT coding scheme is now also available for
the SCT coding scheme (new header file "dcmsr/codes/sct.h").

These new code constants are not yet used within the DCMTK. There will be
another commit replacing the use of the old CODE_SRT_xxx constants by the
respective CODE_SCT_xxx constants. Also the context group and SR template
classes will be updated later on.