Fixed wrong DIMSE status codes (A8xx/A800).
Fixed wrong DIMSE status codes A8xx (for C-STORE and C-FIND) and A800 (for
C-GET and C-MOVE), which were never defined in the official DICOM standard.
Now, the DCMTK uses the correct DIMSE status code 0122H for "SOP Class not
supported" for all DIMSE messages. This bug was contained in the DCMTK since
1993, i.e. already in its first release.

Currently, the DICOM standard is not consistent regarding the error class
(refused, failed or error) for this and other status codes, so in the future
there might be name changes of the corresponding constants/defines.