Notify about association termination.
Notify about the termination of an association using the related method
in the DcmSCP class.  The notify method has been called in DcmSCP but not
on the threaded version of the class (DcmThreadSCP) which has been fixed
with this commit.

Thanks to Brian Wise <brian.wise@medtronic.com> for the report.