Added wide char support to XML parser (Windows).
Added macro that enables the wide char (wchar_t*) support of the XML parser
that is part of the DCMTK (ofstd/ofxml). This support is limited to Windows
systems and still regarded as experimental (see documentation for details).

Instead of defining UNICODE/_UNICODE (which is still not supported by the
DCMTK) the macro WIDE_CHAR_XML_PARSER has been introduced for this purpose.