Enable setting of individual JPEG-LS encoding parameters.
Previously, modifying the JPEG-LS encoding parameters T1, T2, T3 and RESET
required that all parameters be set manually. Now they can be set individually
and the values of the other parameters are computed automatically if needed.
Support for the JPEG-LS parameter LIMIT, which is not accessible through a
public API in CharLS, was removed.