Fixed custom RESET value issue in JPEG-LS encoder.
The JPEG-LS standard allows encoding images with a custom RESET value in
the JPEG-LS preset coding parameters. Due to a bug in CharLS, the use of
custom RESET values caused invalid JPEG-LS bitstreams to be created when
compressing images with more than 8 bits/sample. This bug was registered
as issue #35 in the bug tracker of the CharLS project. Fixed the issue by
back-porting the related fix (commit bd8cb5c) from CharLS 1.1.