Fixed bug in the near-lossless JPEG-LS encoder.
Fixed a bug in the JPEG-LS cooked encoder affecting images with BitsStored
values other than 8 or 16 when compressed in near-lossless mode. The bug caused
the number of bits per pixel fed to the encoder to be always set to 8 or 16,
which could cause pixel values to overflow the permitted value range.