Workaround for ICU on the IBM XL C/C++ compiler.
In some cases the ICU library tries to use the char16_t type even though
the compiler does not understand it. This workaround makes sure, that in
such cases the ICU falls back to uint16_t. Additionally there was a problem
with the definition of U_NOEXCEPT from the internal API of the ICU. This
should also be resolved.