Fixed crash when "tag" attribute is missing.
Fixed crash (e.g. segmentation fault on Linux systems) in xml2dcm when
parsing an XML file where the "tag" attribute is missing for an element.
Also enhanced log output in case the "tag" or "vr" attribute is missing.

Thanks to Sergei Gordey <serg.gordey@gmail.com> for the report and the
samples files (created by a fuzzing tool).