Fixed another issue with wrong VR for sequences.
Fixed another issue that caused a crash in xml2dcm when parsing a "sequence"
element with wrong VR. Made sure that instances of the DcmElement class are
only typecasted if they are of the correct type.

Thanks again to Sergei Gordey <serg.gordey@gmail.com> for the report and
to Maria Nedyak who actually did the analysis and created the samples files.